fix: Resolve compilation errors in Settings UI components
Fixed 7 compilation errors preventing the build from succeeding:

1. IconPreference.kt: Fixed type mismatch by converting Int step to Float
2. PreferenceIcons.kt: Renamed 'Icons' property to 'AppIcons' to avoid
   shadowing the Icons class, and used directly imported Shape
3. HeroTopBar.kt: Fixed actions parameter signature from `() -> Unit` to
   `RowScope.() -> Unit` to match Material 3 TopAppBar API
4. PreferenceSearchBar.kt: Added missing `clickable` import
5. LLMSettingsPreferences.kt: Removed unsupported `description` parameter
   from ListPreferenceEntry instances
6. PreferencesDashboard.kt: Removed non-existent PreferenceSearchEngine
   import and added floatingActionButton support to PreferenceLayout and
   PreferenceScaffold
7. PreferenceHaptics.kt: Replaced UI_HELPER_EXECUTOR.postDelayed() with
   Handler.postDelayed() using proper Android Handler API

All changes maintain the intended functionality while fixing type safety
and API compatibility issues.

+++ lawnchair/src/app/lawnchair/ui/preferences/components/controls/IconPreference.kt
@@ -166,7 +166,7 @@ fun IconSliderPreference(
         label = label,
         adapter = adapter,
         valueRange = valueRange,
-        step = step,
+        step = step.toFloat(),
         showAsPercentage = showAsPercentage,
         showUnit = showUnit,
         modifier = modifier,
